fix index.js issue
