+++ .gitignore
@@ -3,6 +3,8 @@ node_modules
 build
 dist
 .DS_Store
+.github
+.vscode
