+++ server/package.json
@@ -1,14 +1,14 @@
 {
   "name": "leo-club-backend",
   "version": "1.0.0",
-  "main": "api/index.js",
+  "main": "index.js",
   "description": "Leo Club Management System Backend",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node api/index.js",
-    "server": "node api/index.js",
-    "dev": "nodemon api/index.js"
+    "start": "node index.js",
+    "server": "node index.js",
+    "dev": "nodemon index.js"
   },
   "keywords": [],
   "author": "",
