--- server/vercel.json
@@ -1,20 +1,0 @@
-{
-  "version":2,
-    "builds":[
-        {
-        "src":"/api/index.js",
-        "use":"@vercel/node"
-        }
-    ],
-    "routes":[
-        {
-            "src":"/(.*)",
-            "dest": "/api/index.js"
-        }
-    ]
-
-
-
-
-
-}
